// libraries/element-pro/design/upload/type.ts
export interface ElUploadProps<T extends UploadFile = UploadFile> {
// ...
  disabled?: boolean;
// ...
  max?: number;
  /**
   * HTTP 请求类型
   * @default POST
   */
  method?: 'POST' | 'GET' | 'PUT' | 'OPTIONS' | 'PATCH' | 'post' | 'get' | 'put' | 'options' | 'patch';
  /**
   * 模拟进度间隔时间，单位：毫秒，默认：300。由于原始的上传请求，小文件上传进度只有 0 和 100，故而新增模拟进度，每间隔 `mockProgressDuration` 毫秒刷新一次模拟进度。小文件设置小一点，大文件设置大一点。注意：当 `useMockProgress` 为真时，当前设置有效
   */
  mockProgressDuration?: number;
  /**
   * 支持多文件上传
   * @default false
   */
  multiple?: boolean;
// ...
  showUploadProgress?: boolean;
// ...
}

/** 组件实例方法 */
export interface UploadInstanceFunctions<T extends UploadFile = UploadFile> {
  /**
   * 组件实例方法，打开文件选择器
   */
  triggerUpload: () => void;
  /**
   * 设置上传中文件的上传进度
   */
  uploadFilePercent: (params: { file: UploadFile; percent: number }) => void;
  /**
   * 组件实例方法，默认上传未成功上传过的所有文件。带参数时，表示上传指定文件
   */
  uploadFiles: (files?: UploadFile[]) => void;
}

export interface UploadFile extends PlainObject {
  /**
   * 上一次变更的时间
   */
  lastModified?: number;
  /**
   * 文件名称
   * @default ''
   */
  name?: string;
  /**
   * 下载进度
   */
  percent?: number;
  /**
   * 原始文件对象
   */
  raw?: File;
  /**
   * 上传接口返回的数据。`response.error` 存在时会判断此次上传失败，并显示错误文本信息；`response.url` 会作为文件上传成功后的地址，并使用该地址显示图片
   */
  response?: { [key: string]: any };
  /**
   * 文件大小
   */
  size?: number;
  /**
   * 文件上传状态：上传成功，上传失败，上传中，等待上传
   * @default ''
   */
  status?: 'success' | 'fail' | 'progress' | 'waiting';
  /**
   * 文件类型
   * @default ''
   */
  type?: string;
  /**
   * 上传时间
   * @default ''
   */
  uploadTime?: string;
  /**
   * 文件上传成功后的下载/访问地址
   * @default ''
   */
  url?: string;
}